apps/ai/requirements.txt
accelerate
basicsr
customtkinter
demucs
diffusers
faster-whisper
gfpgan
kornia
numpy
opencv-python
Pillow
realesrgan
torch
transformers
